% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufc-data.R
\docType{data}
\name{ufc_stats}
\alias{ufc_stats}
\title{data.frame containing UFC fight Statistics}
\format{An object of class \code{data.frame} with 25044 rows and 37 columns.}
\usage{
data(ufc_stats)
}
\description{
UFC fight statistics data from the very beginning to today.
}
\examples{
data(ufc_stats)
}
\references{
http://www.ufcstats.com/statistics/events/completed
}
\keyword{datasets}
