% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_data.R
\name{refresh_data}
\alias{refresh_data}
\title{Refresh UFC data with the latest fights}
\usage{
refresh_data()
}
\description{
This function updates the rda file contained in the package folder data/ufc_stats.rda.
}
