% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufc-data.R
\docType{data}
\name{ufc_stats}
\alias{ufc_stats}
\title{data.frame containing UFC fight stats}
\format{
An object of class \code{data.frame} with 27668 rows and 36 columns.
}
\usage{
data(ufc_stats)
}
\description{
UFC fight statistics data from the very beginning to today.
}
\examples{
data(ufc_stats)
}
\references{
http://www.ufcstats.com/statistics/events/completed
}
\keyword{datasets}
