% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufc-data.R
\docType{data}
\name{ufc_stats}
\alias{ufc_stats}
\title{UFC Fight Statistics}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(ufc_stats)
}
\description{
Data UFC Fights from the very beginning to today.
}
\examples{
data(ufc_stats)
}
\references{
(\href{http://www.ufcstats.com/statistics/events/completed}{UFC Stats})
}
\keyword{datasets}
